add bdd test
